wo#7616 - refactor handle_next_timer_event() so that it can be unit tested
