+++ programs/pluto/Makefile.depend.linux
@@ -2164,6 +2164,37 @@ rcv_whack.o: rcv_whack.c ../../include/openswan.h \
  ../../include/sha2.h \
  ../../include/kernel_alg.h \
  ../../include/openswan/pfkeyv2.h ike_alg.h
+replace.o: replace.c ../../include/openswan.h \
+ ../../ports/linux/include/sysdep.h \
+ ../../include/sysqueue.h \
+ ../../include/constants.h \
+ ../../include/biglset.h \
+ ../../include/openswan/passert.h \
+ ../../include/ietf_constants.h \
+ ../../include/pluto_constants.h \
+ ../../include/names_constant.h \
+ ../../include/pluto/defs.h \
+ ../../include/oswalloc.h \
+ ../../include/constants.h ../../include/id.h \
+ ../../include/pluto/connections.h \
+ ../../include/x509.h ../../include/pgp.h \
+ ../../include/certs.h \
+ ../../include/openswan/ipsec_policy.h \
+ ../../include/secrets.h ../../include/id.h \
+ ../../include/x509.h ../../include/pgp.h \
+ ../../include/pluto/defs.h \
+ ../../include/pluto/log.h \
+ ../../include/oswlog.h \
+ ../../include/openswan/passert.h state.h \
+ ../../include/pluto/quirks.h \
+ ../../include/packet.h \
+ ../../include/pluto/demux.h \
+ ../../include/pluto/server.h \
+ ../../include/pluto/quirks.h ipsec_doi.h ikev1_quick.h \
+ timer.h ../../include/oswtime.h \
+ ../../include/pluto/dpd.h replace.h \
+ ../../include/pluto/nat_traversal.h \
+ ../../include/pluto/demux.h
 security_selinux.o: security_selinux.c
 server.o: server.c ../../include/openswan.h \
  ../../ports/linux/include/sysdep.h \
